fix: fix review scheme, ranked title and review style
