+++ .eslintrc.js
@@ -39,7 +39,8 @@ module.exports = {
         "jsx-a11y/click-events-have-key-events": "off",
         "jsx-a11y/no-static-element-interactions": "off",
         "jsx-a11y/label-has-associated-control": "off",
-        "prettier/prettier": "error",
+        // "prettier/prettier": "error",
+        "prettier/prettier": "off",
         "no-unused-vars": [
           "error",
           {
