+++ .vscode/settings.json
@@ -15,7 +15,7 @@
   "phpsab.standard": "WordPress",
   "[php]": {
     "breadcrumbs.showConstructors": true,
-    "editor.defaultFormatter": "valeryanm.vscode-phpsab"
+    "editor.defaultFormatter": "bmewburn.vscode-intelephense-client"
     // "editor.defaultFormatter": "bmewburn.vscode-intelephense-client"
   },
