+++ package.json
@@ -79,8 +79,6 @@
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.2",
     "cssnano": "^5.1.14",
-    "nanoid": "^4.0.0",
-    "picocolors": "^1.0.0",
     "esbuild": "^0.15.14",
     "eslint": "^8.27.0",
     "eslint-config-prettier": "^8.5.0",
@@ -97,10 +95,13 @@
     "gulp-sass": "^5.1.0",
     "gulp-terser": "^2.1.0",
     "husky": "^8.0.2",
-    "inquirer": "^9.1.4",
+    "inquirer": "8.0.1",
+    "jest": "^29.3.1",
     "lint-staged": "^13.0.3",
     "merge-stream": "^2.0.0",
+    "nanoid": "^4.0.0",
     "node-sass": "^8.0.0",
+    "picocolors": "^1.0.0",
     "postcss": "^8.4.19",
     "postcss-import": "^15.0.0",
     "postcss-loader": "^7.0.1",
@@ -108,6 +109,7 @@
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
     "process": "^0.11.10",
+    "puppeteer-core": "^19.2.2",
     "react-icons": "^4.6.0",
     "rimraf": "^3.0.2",
     "sass": "^1.56.1",
