+++ src/scss/components/_review.scss
@@ -1,0 +1,267 @@
+/* Review */
+.review_average_stars {
+    float: right;
+  }
+  .review_entry,
+  .review_percentage_entry {
+    margin-right: auto;
+  }
+  .review_entry .components-icon-button,
+  .review_percentage_entry .components-icon-button {
+    padding: 0;
+  }
+  .review_entry .dashicons-trash,
+  .review_percentage_entry .dashicons-trash {
+    color: #aaa;
+    float: right;
+  }
+  .review_entry .dashicons-trash:hover,
+  .review_percentage_entry .dashicons-trash:hover {
+    color: #222;
+  }
+  .review_block .review_image_container {
+    position: relative;
+  }
+  .review_block .review_image_container .remove-image {
+    position: absolute;
+    right: 0;
+  }
+  .review_block .review_image_container .review_image {
+    height: auto;
+    width: auto;
+  }
+  .review_description_container .review_description {
+    margin-top: 0;
+    margin-bottom: 0;
+    padding-top: 0;
+    padding-bottom: 0;
+  }
+  .review_summary .review_overall_value {
+    display: flex;
+  }
+  .review_summary .review_overall_value > .rich-text {
+    flex-basis: 75%;
+  }
+  .review_summary .review_overall_value > .review_average {
+    flex-basis: 25%;
+    margin-left: auto;
+  }
+  .review_cuisine_list {
+    list-style: none;
+  }
+  .review_cuisine_list li {
+    display: inline;
+    border: 1px solid #000;
+    border-radius: 5px;
+    padding: 3px;
+  }
+  #review_event_page_input input,
+  #review_video_url_input input,
+  #review_item_page_input input {
+    width: 200px;
+  }
+  
+  /* Review */
+  .review_block {
+    padding: 25px 40px 35px 40px;
+    box-shadow: 0px 0px 4px 1px #d6d6d6;
+    border-radius: 5px;
+    margin: 30px auto;
+  }
+  .review_block svg:not(:root) {
+    margin: 0 1px;
+  }
+  .review_block svg:not(:root):first-child {
+    margin-left: 0;
+  }
+  .review_block svg:not(:root):last-child {
+    margin-right: 0;
+  }
+  .review_block .review_image {
+    margin-top: 10px;
+    float: right;
+    object-fit: contain;
+  }
+  .review_overall_value {
+    display: grid;
+    grid-template-columns: 75% 25%;
+  }
+  .review_entry,
+  .review_percentage_entry {
+    padding: 10px 0 10px 0;
+    border-bottom: 1px solid #e9e9e9;
+  }
+  .review_entry {
+    display: flex;
+  }
+  .review_percentage_entry {
+    display: grid;
+    grid-template-columns: 25% 75%;
+  }
+  .review_rating {
+    text-align: right;
+    font-size: 40px;
+  }
+  .review_average {
+    display: flex;
+    flex-direction: column;
+  }
+  .review_average_stars {
+    display: flex;
+    grid-area: auto;
+    justify-self: self-end;
+    height: 50px;
+    line-height: 60px;
+    margin-left: auto;
+    align-items: center;
+  }
+  .review_cta_panel {
+    display: grid;
+    grid-template-columns: 2fr auto;
+    grid-template-areas: 'cta . stars';
+  }
+  .review_add_entry {
+    left: 50%;
+    width: 100%;
+  }
+  .review_link_input {
+    display: flex;
+    flex-direction: row;
+  }
+  .review_cta_btn {
+    min-width: 100px;
+    padding: 15px 10px;
+    border-radius: 6px;
+    display: inline-block;
+    border-width: 1px;
+    border-style: solid;
+    text-align: center;
+    text-transform: none;
+    cursor: pointer;
+    font-weight: normal;
+    background-image: none;
+  }
+  .review_cta_btn:hover,
+  .review_cta_btn:active {
+    background-image: none;
+  }
+  .review_item_name {
+    font-size: 28px !important;
+    margin: 0 !important;
+    font-weight: 600 !important;
+  }
+  .review_summary_title {
+    margin: 23px 0 13px 0 !important;
+    font-size: 24px !important;
+    font-weight: 600 !important;
+  }
+  .review_stars {
+    display: flex;
+    margin-left: auto;
+    padding-top: 3px;
+  }
+  .review_description {
+    padding: 0;
+  }
+  .review_description_container {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 10px;
+    overflow: auto;
+  }
+  .review_left_image {
+    flex-direction: row;
+  }
+  .review_right_image {
+    flex-direction: row-reverse;
+  }
+  .review_top_image {
+    flex-direction: column;
+  }
+  .review_bottom_image {
+    flex-direction: column-reverse;
+  }
+  .review_percentage {
+    display: grid;
+    grid-template-columns: 4fr 1fr;
+    overflow: auto;
+    margin-left: auto;
+  }
+  .review_percentage > svg {
+    height: 10px;
+  }
+  .review_percentage > div {
+    text-align: end;
+  }
+  .review_percentage_bar {
+    width: 100%;
+    align-self: center;
+  }
+  .review_percentage_bar_trail {
+    fill-opacity: 0;
+    stroke-linecap: round;
+  }
+  .review_percentage_bar_path {
+    stroke-dasharray: 100px, 100px;
+    transition: stroke-dashoffset 0.3s ease 0s, stroke 0.3s linear 0s;
+    fill-opacity: 0;
+    stroke-linecap: round;
+  }
+  .review_cta_main {
+    display: flex;
+  }
+  @media (max-width: 7688px) {
+    .review_block {
+      padding: 20px 15px;
+    }
+  }
+  @media (max-width: 576px) {
+    .review_cta_btn {
+      padding: 15px 10px;
+    }
+    .review_image {
+      float: right;
+    }
+  }
+  @media (max-width: 480px) {
+    .review_rating {
+      font-size: 20px;
+      font-weight: 600;
+    }
+    .review_entry {
+      font-size: 18px;
+    }
+    .review_cta_btn {
+      padding: 15px 8px;
+    }
+  }
+  @media (max-width: 460px) {
+    .review_average_stars svg {
+      width: 12px;
+      height: 12px;
+    }
+  }
+  @media (max-width: 375px) {
+    .review_block {
+      padding: 20px 15px;
+    }
+    .review_block button > svg {
+      width: 12px;
+      height: 12px;
+    }
+  }
+  @media (max-width: 320px) {
+    .review_block {
+      padding: 20px 10px;
+    }
+    .review_block button > svg {
+      width: 8px;
+      height: 8px;
+      margin: 0;
+    }
+    .review_average_stars {
+      height: 45px;
+      line-height: 40px;
+      margin-left: auto;
+    }
+  }
