+++ src/scss/frontend.scss
@@ -1,1 +1,3 @@
 @import "./main";
+
+@import "components/review";
